+++ apps/modernization-ui/src/apps/patient/data/name/asName.spec.ts
@@ -1,4 +1,4 @@
-import { asName } from '..';
+import { asName } from './asName';
 
 describe('when mapping a name entry to a format accepted by the API', () => {
     it('should include the as of date', () => {
