+++ apps/modernization-ui/src/apps/patient/data/name/index.ts
@@ -1,0 +1,5 @@
+export { initial } from './entry';
+export type { NameEntry } from './entry';
+
+export { NameEntryFields } from './NameEntryFields';
+export { asName } from './asName';
